+++ views/breads/submitBread.jsx
@@ -1,5 +1,5 @@
 const React = require('react')
-const DefaultLayout = require("./layouts/default")
+const DefaultLayout = require("../layouts/default")
 
 function submitBread() {
     return (
